gears.geometry: Add area_remove
